Stratum_index changes to xvw
SImplifying logic of onevax_xvw and putting lots of the wrangling into a stratum_index file.

Also N.B. doing so I've noticed a couple of issues with the current implementation of vaccination strategies other than xpvwrh if n_diag_rec > 1. Will be fixed in upcoming commits (to be labelled stratum_index changes to ....)